[ADD]GMO: new background colours have been introduced to distinguish between data types accepted in a given cell: Monetary cell background: white (normal) Percent cell background: light yellow Integer cell background: light orange Text cell background: light violet Date cell background: light green
Moreover, the navigation in the TkTable widget has been improved:
With the tabulator key, you go to the cell on the right of the current cell.
With the return key, you go to the cell at the bottom of the current cell.

Unfortunately, for cells containing a drop-down list, the selection is not visible, but it is effective.